c19-backend/requirements.txt
Django
psycopg2
djangorestframework
requests
pydevd-pycharm~=193.6494.30
